+++ dapps/Cargo.toml
@@ -22,7 +22,6 @@ serde_json = "0.7.0"
 serde_macros = { version = "0.7.0", optional = true }
 zip = { version = "0.1", default-features = false }
 ethabi = "0.2.1"
-tiny-keccak= "1.0"
 ethcore-rpc = { path = "../rpc" }
 ethcore-util = { path = "../util" }
 ethcore-devtools = { path = "../devtools" }
