+++ dapps/src/apps/fetcher.rs
@@ -19,7 +19,6 @@
 //! Uses `URLHint` to resolve addresses into Dapps bundle file location.
 
 use zip;
-use tiny_keccak::Keccak;
 use std::{fs, env};
 use std::io::{self, Read, Write};
 use std::path::PathBuf;
@@ -32,6 +31,7 @@ use hyper::status::StatusCode;
 
 use random_filename;
 use util::{Mutex, H256};
+use util::sha3::sha3;
 use page::LocalPageEndpoint;
 use handlers::{ContentHandler, AppFetcherHandler, DappHandler};
 use endpoint::{Endpoint, EndpointPath, Handler};
@@ -156,45 +156,6 @@ impl From<zip::result::ZipError> for ValidationError {
 	fn from(err: zip::result::ZipError) -> Self {
 		ValidationError::Zip(err)
 	}
-}
-
-
-fn sha3(file: &mut fs::File) -> Result<H256, io::Error> {
-	let mut output = [0u8; 32];
-	let mut input = [0u8; 1024];
-	let mut sha3 = Keccak::new_keccak256();
-
-	// read file
-	loop {
-		let some = try!(file.read(&mut input));
-		if some == 0 {
-			break;
-		}
-		sha3.update(&input[0..some]);
-	}
-
-	sha3.finalize(&mut output);
-	Ok(output.into())
-}
-
-#[cfg(test)]
-#[test]
-fn should_sha3_a_file() {
-	// given
-	use ethcore_devtools::RandomTempPath;
-	let path = RandomTempPath::new();
-	// Prepare file
-	{
-		let mut file = fs::File::create(&path).unwrap();
-		file.write_all(b"something").unwrap();
-	}
-
-	let mut file = fs::File::open(&path).unwrap();
-	// when
-	let hash = sha3(&mut file).unwrap();
-
-	// then
-	assert_eq!(format!("{:?}", hash), "68371d7e884c168ae2022c82bd837d51837718a7f7dfb7aa3f753074a35e1d87");
 }
 
 struct DappInstaller {
