+++ util/Cargo.toml
@@ -32,6 +32,7 @@ parking_lot = "0.2.6"
 using_queue = { path = "using_queue" }
 table = { path = "table" }
 ansi_term = "0.7"
+tiny-keccak= "1.0"
 
 [features]
 default = []
